pyproject.toml
[project]
name = "rescue-dog-aggregator"
version = "1.0.0"
description = "Open-source platform aggregating rescue dogs from multiple organizations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "uvicorn[standard]>=0.38.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.44,<3.0.0",
    "alembic>=1.17.2,<2.0.0",
    "starlette>=0.50.0",
    "urllib3>=2.6.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "jsonschema>=4.25.1",
    "sentry-sdk[fastapi,sqlalchemy]>=2.47.0",
    "werkzeug>=3.1.5",
    "mcp>=1.23.1",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
    "playwright>=1.49.0",
    "nest-asyncio>=1.6.0",
    "pillow>=12.0.0",
    "click>=8.3.1",
    "colorama>=0.4.6",
    "rich>=14.2.0",
    "langdetect==1.0.9",
    "python-dotenv>=1.2.1",
    "boto3>=1.42.13",
    "python-dateutil>=2.9.0",
    "tenacity>=9.1.2",
    "iniconfig>=2.3.0",
    "cachetools>=6.2.4",
    "pytest-benchmark>=5.2.3",
    "firecrawl-py>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.14.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "ipython>=8.0.0",
    "watchdog>=3.0.0",
    "cryptography>=41.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "scrapers", "services", "utils", "database"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.14.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "ipython>=8.0.0",
    "watchdog>=3.0.0",
    "cryptography>=41.0.0",
]

[tool.ruff]
line-length = 200
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "venv",
    ".env",
    "env",
    "build",
    "dist",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    ".coverage",
    "htmlcov",
    "node_modules",
    "frontend",
    "migrations",
]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "E402",   # Module level import not at top (often intentional)
    "E712",   # Comparison to True/False (style preference)
    "E722",   # Bare except (needs case-by-case review)
    "F401",   # Unused import (fix in follow-up)
    "F402",   # Import shadowed by loop variable (fix in follow-up)
    "F601",   # Dict key repeated (fix in follow-up)
    "F811",   # Redefinition of unused name (fix in follow-up)
    "F821",   # Undefined name (fix in follow-up)
    "F841",   # Unused variable (often intentional in tests)
    "W291",   # Trailing whitespace
    "W293",   # Blank line contains whitespace
]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
known-first-party = ["api", "scrapers", "services", "utils", "database", "tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
